+++ src/toml.rs
@@ -1,4 +1,6 @@
-use crate::types::DataFormat;
+use crate::common::*;
+use crate::path::*;
+use crate::types::{DataFormat, ToStrError};
 use crate::value::*;
 use std::collections::HashMap;
 use toml::Value as TomlValue;
@@ -9,8 +11,15 @@ impl DataFormat for Toml {
     fn from_str(&self, s: &str) -> Option<Value> {
         to_value(&toml::from_str(s).ok()?)
     }
-    fn to_str(&self, value: &Value, _: bool) -> Option<String> {
-        toml::to_string(&to_toml_value(value).unwrap()).ok()
+    fn to_str(&self, value: &Value, _: bool) -> Result<String, ToStrError> {
+        Ok(
+            toml::to_string(&to_toml_value(value, &[]).map_err(|err| match err {
+                ToTomlValueError::UnsupportedType((value, path)) => {
+                    ToStrError::UnsupportedType((value.type_encoded(), path))
+                }
+            })?)
+            .map_err(to_parse_error)?,
+        )
     }
 }
 
@@ -55,29 +64,66 @@ fn to_value(value: &TomlValue) -> Option<Value> {
     None
 }
 
-fn to_toml_value(value: &Value) -> Option<TomlValue> {
+#[derive(Debug)]
+enum ToTomlValueError {
+    UnsupportedType((Value, Vec<PathEntry>)),
+}
+
+fn to_toml_value(value: &Value, path: &[PathEntry]) -> Result<TomlValue, ToTomlValueError> {
     match value {
-        Value::String(value) => Some(TomlValue::from(value.to_owned())),
-        Value::Int(value) => Some(TomlValue::from(*value)),
-        Value::Float(value) => Some(TomlValue::from(*value)),
-        Value::Bool(value) => Some(TomlValue::from(*value)),
-        Value::List(list) => Some(TomlValue::Array(
+        Value::String(value) => Ok(TomlValue::from(value.to_owned())),
+        Value::Int(value) => Ok(TomlValue::from(*value)),
+        Value::Float(value) => Ok(TomlValue::from(*value)),
+        Value::Bool(value) => Ok(TomlValue::from(*value)),
+        Value::List(list) => Ok(TomlValue::Array(
             list.iter()
-                .map(|value| to_toml_value(value).unwrap())
-                .collect(),
+                .enumerate()
+                .map(|(i, value)| {
+                    to_toml_value(value, &{
+                        let mut path_new = path.to_owned();
+                        path_new.push(PathEntry::Index(i));
+
+                        path_new
+                    })
+                })
+                .collect::<Result<Vec<TomlValue>, ToTomlValueError>>()?,
         )),
         Value::Object(value) => {
             let mut obj = toml::Table::new();
 
             for (key, value) in value {
+                if let Value::Null = value {
+                    return Err(ToTomlValueError::UnsupportedType((Value::Null, {
+                        let mut path_new = path.to_owned();
+                        path_new.push(PathEntry::Field(if let Key::String(s) = key {
+                            s.to_owned()
+                        } else {
+                            panic!("Failed to parse key");
+                        }));
+
+                        path_new
+                    })));
+                }
+
                 obj.insert(
                     key_to_string(key),
-                    to_toml_value(value).unwrap_or(TomlValue::from("null")),
+                    to_toml_value(value, &{
+                        let mut path_new = path.to_owned();
+                        path_new.push(match key {
+                            Key::String(s) => PathEntry::Field(s.to_owned()),
+                            Key::Int(i) => PathEntry::Index((*i).try_into().unwrap()),
+                        });
+
+                        path_new
+                    })?,
                 );
             }
 
-            Some(TomlValue::from(obj))
+            Ok(TomlValue::from(obj))
         }
-        Value::Null => Some(TomlValue::String("null".to_string())),
+        Value::Null => Err(ToTomlValueError::UnsupportedType((
+            Value::Null,
+            path.to_vec(),
+        ))),
     }
 }